+++ examples/shapes_triangle_intersect.rs
@@ -78,7 +78,7 @@ fn main() {
     let r: Ray = Ray {
         o: o,
         d: d,
-        t_max: std::f64::INFINITY,
+        t_max: std::f32::INFINITY,
         time: 0.0,
         differential: None,
     };
@@ -96,7 +96,7 @@ fn main() {
     let r: Ray = Ray {
         o: o,
         d: d,
-        t_max: std::f64::INFINITY,
+        t_max: std::f32::INFINITY,
         time: 0.0,
         differential: None,
     };
